% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentCurv.R
\name{CentCurv}
\alias{CentCurv}
\title{Centile curves based on different methods}
\usage{
CentCurv(
  x,
  y,
  window.select = "CV",
  method,
  lambdas = 0,
  kernel = "gauss",
  cents = c(0.03, 0.25, 0.5, 0.75, 0.97),
  disp_window = FALSE
)
}
\arguments{
\item{x}{the explanatory variable - numeric}

\item{y}{the response variable - numeric}

\item{window.select}{the bandwidth method choice: CV or plug-in. Default is CV.}

\item{method}{str - The method choosen for displaying the curve. Could be: Local constant, Local linear, Cubic splines or B-splines.}

\item{lambdas}{set to 0. To be set if method chosen is "B-Splines". Tunes the tradeoff between the goodness of fit and the regularity of the spline - numeric value or numeric vector}

\item{cents}{A numeric vector that represents the centiles calculated. Default is set to cents=c(0.03,0.25,0.5,0.75,0.97) )}
}
\value{
Plots centile curves according to the method chosen
}
\description{
Centile curves based on different methods
}
\examples{
#create a sample data frame
sample<-data.frame(sample(30:42,50,rep=TRUE),sample(800:5000,50,rep=TRUE))
colnames(sample)<-c("Gestational Age in weeks","Weight in gramms")
x<-sample$`Gestational Age in weeks`
y<-sample$`Weight in gramms`
CentCurv(x,y,method='Cubic splines')
CentCurv(x,y,method='B-Splines',lambdas=1)


}
