% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCurv.R
\name{compareCurv}
\alias{compareCurv}
\title{Centile curves according to different methods}
\usage{
compareCurv(x, y, bandwidth.method = "CV", lambdas, data)
}
\arguments{
\item{x}{the explanatory variable - numeric}

\item{y}{the response variable - numeric}

\item{bandwidth.method}{the method chosen to calculate bandwdth. Could be cross validation or Plug-in. Default is set to CV.}

\item{lambdas}{to be set for "B-Splines". Can be a vector or a single numeric value. Tunes the tradeoff between the goodness of fit and the regularity of the spline - numeric value or numeric vector}

\item{data}{the abnormal/external data we want to compare the curves with}
}
\value{
Plots centile curves with the different methods and displays them on the same figure as the noise data to compare
}
\description{
Centile curves according to different methods
}
