% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCurv.R
\name{cubicSpline0}
\alias{cubicSpline0}
\title{Centile curves using cubic splines compared to noise data}
\usage{
cubicSpline0(x, y, cents = c(0.03, 0.25, 0.5, 0.75, 0.97), data)
}
\arguments{
\item{x}{the explanatory variable - numeric}

\item{y}{the response variable - numeric}

\item{cents}{A numeric vector that represents the centiles calculated. Default is set to cents=c(0.03,0.25,0.5,0.75,0.97) )}

\item{data}{the noise data to be compared to}
}
\value{
Plots centile curves with cubic splines and displays them on the same figure as the noise data
}
\description{
Centile curves using cubic splines compared to noise data
}
\examples{
#create a sample data frame
sample<-data.frame(sample(30:42,50,rep=TRUE),sample(800:5000,50,rep=TRUE))
colnames(sample)<-c("Gestational Age in weeks","Weight in gramms")
x<-sample$`Gestational Age in weeks`
y<-sample$`Weight in gramms`
abnormal<-data.frame(sample(30:42,10,rep=TRUE),sample(800:5000,10,rep=TRUE))
colnames(abnormal)<-c("Gestational Age in weeks","Weight in gramms")
cubicSpline0(x,y,cents=c(0.03,0.25,0.5,0.75,0.97),data=abnormal)

}
