% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelFonctions.R
\name{kernel.select}
\alias{kernel.select}
\title{Kernel selection}
\usage{
kernel.select(kernel)
}
\arguments{
\item{kernel}{a string - the kernel chosen. Can be trig, gauss, circ, cubic or epan.}
}
\value{
the function of the Kernel selected
}
\description{
Kernel selection
}
\examples{
kernel.select("cubic")
}
