% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localCst.R
\name{localCst}
\alias{localCst}
\title{Local constant estimator}
\usage{
localCst(x, y, bandwidth.method = "CV", cents = c(0.03, 0.25, 0.5, 0.75, 0.97))
}
\arguments{
\item{x}{the explanatory variable - numeric}

\item{y}{the response variable - numeric}

\item{bandwidth.method}{the bandwidth method choice: CV or plug-in. Default is CV.}

\item{cents}{A numeric vector that represents the centiles calculated. Default is set to cents=c(0.03,0.25,0.5,0.75,0.97) )}
}
\value{
Plots the centile curves and returns a list object containing bandwidth value and estimated centiles values. Kernel used is gaussian.
}
\description{
Local constant estimator
}
\examples{
#create an example data frame
example<-data.frame(sample(30:42,50,rep=TRUE),sample(800:5000,50,rep=TRUE))
colnames(example)<-c("Gestational Age in weeks","Weight in gramms")
x<-example$`Gestational Age in weeks`
y<-example$`Weight in gramms`
#calculate the centile and plot the curves
localCst(x,y)

}
