% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localLin.R
\name{localLin}
\alias{localLin}
\title{Local linear estimator}
\usage{
localLin(
  x,
  y,
  bandwidth.method = "CV",
  kernel = "gauss",
  cents = c(0.03, 0.25, 0.5, 0.75, 0.97),
  disp_window = TRUE
)
}
\arguments{
\item{x}{the explanatory variable - numeric}

\item{y}{the response variable - numeric}

\item{bandwidth.method}{the bandwidth method choice: CV or plug-in. Default is CV.}

\item{cents}{A numeric vector that represents the centiles calculated. Default is set to cents=c(0.03,0.25,0.5,0.75,0.97) )}
}
\value{
Plots the centile curves and returns a list object containing bandwidth value and estimated centiles values. Kernel used is gaussian.
}
\description{
Local linear estimator
}
\examples{
#create a data frame
example<-data.frame(sample(30:42,50,rep=TRUE),sample(800:5000,50,rep=TRUE))
colnames(example)<-c("Gestational Age in weeks","Weight in gramms")
x<-example$`Gestational Age in weeks`
y<-example$`Weight in gramms`
#calculate the centiles and plot the curves
localLin(x,y)

}
