% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moyenne.R
\name{moyenne}
\alias{moyenne}
\title{Title
Fonction faisant la moyenne d'un vecteur}
\usage{
moyenne(x)
}
\arguments{
\item{x}{vecteur numérique}
}
\value{
moyenne du vecteur
}
\description{
Title
Fonction faisant la moyenne d'un vecteur
}
\examples{
moyenne(c(1, 2, 3, 4))
}
