% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegSpline.R
\name{RegSpline}
\alias{RegSpline}
\title{Regression Splines regression to draw centile curves}
\usage{
RegSpline(x, y, cents = c(0.03, 0.25, 0.5, 0.75, 0.97))
}
\arguments{
\item{x}{the explanatory variable - numeric}

\item{y}{the response variable - numeric}

\item{cents}{A numeric vector that represents the centiles calculated. Default is set to cents=c(0.03,0.25,0.5,0.75,0.97) )}
}
\value{
Plots the centile curves
}
\description{
Regression Splines regression to draw centile curves
}
\examples{
#create a data frame
example<-data.frame(sample(30:42,100,rep=TRUE),sample(800:5000,100,rep=TRUE))
colnames(example)<-c("Gestational Age in weeks","Weight in gramms")
x<-example$`Gestational Age in weeks`
y<-example$`Weight in gramms`
#calculate the centiles and plot the curves
RegSpline(x,y)


}
