% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Test
Fonction appelant la fonction faisant la moyenne d'un vecteur}
\usage{
test(x)
}
\arguments{
\item{x}{vecteur numérique}
}
\value{
moyenne du vecteur
}
\description{
Test
Fonction appelant la fonction faisant la moyenne d'un vecteur
}
\examples{
test(c(1, 2, 3, 4))
}
