% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{window}
\alias{window}
\title{Window selection function}
\usage{
window(x, y, method = "CV")
}
\arguments{
\item{x}{the explanatory variable - numeric}

\item{y}{the response variable - numeric}

\item{method}{the bandwidth method choice: CV or plug-in. Default is CV.}
}
\value{
Calculates the window value using cross validation or plug-in method
}
\description{
Window selection function
}
\examples{
#create a data frame
example<-data.frame(sample(30:42,10,rep=TRUE),sample(800:5000,10,rep=TRUE))
colnames(example)<-c("Gestational Age in weeks","Weight in gramms")
x<-example$`Gestational Age in weeks`
y<-example$`Weight in gramms`
#calculate the window value
window(x,y)

}
